add Card comp to CardGroup
