+++ src/lib/components/card/Card.scss
@@ -1,0 +1,50 @@
+.flex-col {
+    display: flex;
+    flex-direction: column;
+}
+
+.card {
+    background-color: var(--js-yellow);
+    border-radius: 10px;
+    border: transparent solid 3px;
+    padding: 1rem;
+    cursor: pointer;
+    transition: transform .2s;
+    width: auto;
+    max-width: 40ch;
+}
+
+@media(min-width: 1080px) {
+    .card {
+        height: 10rem;
+        width: 17rem;
+        padding: 25px;
+    }
+}
+
+.card:hover {
+    box-shadow: 0 8px 8px rgba(124, 124, 111, 0.25);
+    border: #d4d1d1 solid 3px;
+    background-color: black;
+    /* transform: scale(1.05); */
+}
+
+.card:hover > h3 {
+    color: #d4d1d1;
+    text-decoration: underline;
+}
+
+.card:hover > p {
+    color: #d4d1d1;
+}
+
+.card h3 {
+    color: var(--js-grey);
+    font-size: var(--step-0);
+    margin-bottom: 0.25em;
+}
+
+.card p {
+    font-size: var(--step--1);
+    color: var(--js-grey)
+}
