+++ src/lib/components/cardGroup/CardGroup.scss
@@ -1,0 +1,34 @@
+.cards {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    gap: 20px;
+    padding: 30px;
+}
+
+a {
+    text-decoration: none;
+}
+
+/* Styles for each category */
+.category {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    border: 1px solid #ddd;
+    border-radius: 8px;
+    padding: 20px;
+    /* background-color: #fff; */
+}
+
+.category h2 {
+    margin-top: 0;
+    font-size: 1.5em;
+}
+
+.category__p {
+    margin: 0;
+    color: #d4d1d1;
+    max-width: 60ch;
+}
+
