+++ src/pages/arrays/ArrayPage.css
@@ -4,6 +4,11 @@ img {
     width: 50%;
 }
 
+/* img {
+    max-width: 100%;
+    height: auto;
+} */
+
 header {
     background: var(--header);
     width: 100%;
@@ -40,94 +45,4 @@ header {
     gap: 30px;
     width: 90%;
     margin-inline: auto;
-}
-
-.cards {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    gap: 20px;
-    padding: 30px;
-}
-
-a {
-    text-decoration: none;
-}
-
-.flex-col {
-    display: flex;
-    flex-direction: column;
-}
-
-.card {
-    background-color: var(--js-yellow);
-    border-radius: 10px;
-    border: transparent solid 3px;
-    padding: 1rem;
-    cursor: pointer;
-    transition: transform .2s;
-    width: auto;
-    max-width: 40ch;
-}
-
-@media(min-width: 1080px) {
-    .card {
-        height: 10rem;
-        width: 17rem;
-        padding: 25px;
-    }
-}
-
-.card:hover {
-    box-shadow: 0 8px 8px rgba(124, 124, 111, 0.25);
-    border: #d4d1d1 solid 3px;
-    background-color: black;
-    /* transform: scale(1.05); */
-}
-
-.card:hover > h3 {
-    color: #d4d1d1;
-    text-decoration: underline;
-}
-
-.card:hover > p {
-    color: #d4d1d1;
-}
-
-img {
-    max-width: 100%;
-    height: auto;
-}
-
-/* Styles for each category */
-.category {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    border: 1px solid #ddd;
-    border-radius: 8px;
-    padding: 20px;
-    /* background-color: #fff; */
-}
-
-.category h2 {
-    margin-top: 0;
-    font-size: 1.5em;
-}
-
-.category p {
-    margin: 0;
-    color: #d4d1d1;
-    max-width: 60ch;
-}
-
-.card h3 {
-    color: var(--js-grey);
-    font-size: var(--step-0);
-    margin-bottom: 0.25em;
-}
-
-.card p {
-    font-size: var(--step--1);
-    color: var(--js-grey)
 }
